Move `DefaultStateRegistry` to the package level.
It makes its existence more obvious. Also, removed validation of state in the `OrderAggregate`. It's too late to validate anything at the stage of applying event data. It should be done in a command handler instead.
